pig-video-detector/requirements.txt
flask==3.0.2
ultralytics==8.1.28
supervision==0.19.0
numpy>=1.26.4
opencv-python-headless>=4.9.0.80
python-dotenv==1.0.1
gunicorn==21.2.0
Pillow>=10.0.0
torch>=2.1.0